bts: Lower RT_UTIL to 39 to fix display underruns on zuma
Display underruns are observed sometimes, with the minimum MIF frequency
requested by decon being 414690 kHz, which selects the lowest MIF
frequency step (421000 kHz). This is too slow and results in underruns.

Changing RT_UTIL to 39 makes the requested frequency become 425323 kHz
instead, which is >421000 kHz and thus selects the next MIF frequency step
above, resolving the display underrun. This particular underrun is observed
while pinching to zoom in Google Photos with a 50 MP photo captured of a
complex outdoor scene, with the underrun only occurring while viewing the
photo fullscreen such that it fills the entire display.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>