bts: Lower RT_UTIL to 30 to mitigate display underruns on zuma
Display underruns are observed sometimes, with the minimum MIF frequency
requested by decon being 414690 kHz, which selects the lowest MIF
frequency step (421000 kHz). This is too slow and results in underruns.

Changing RT_UTIL to 39 makes the requested frequency become 425323 kHz
instead, which is >421000 kHz and thus selects the next MIF frequency step
above, resolving the display underrun. This particular underrun is observed
while pinching to zoom in Google Photos with a HEIC photo captured from an
iPhone 12 Pro, with the underrun only occurring while viewing the photo
fullscreen.

That being said, there are still occasional display underruns which cannot
be easily reproduced, some of which randomly occur with specific games or
sequences of events.

Since a RT_UTIL of 39 isn't enough to zap all the bogey underruns, crank it
down to 30 which effectively selects the next higher up MIF frequency step
when decon's bandwidth needs are low. This doesn't incur any significant
energy usage increase and appears to provide a sufficient guard band for
mitigating the display underruns on zuma.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>