cal-if: Use the correct DT bindings based on the SoC
The gs101 bindings are always used when they shouldn't be. Fix it so that
the correct bindings are included for gs201 and zuma. As luck would have
it, this shouldn't result in any functional change since the gs101 macros
that are used are defined with the same values on gs201 and zuma.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>