tensor_aio: Use the system timer directly for more precise timekeeping
On Tensor SoCs, the generic system timer increments once every ~40.69 ns
(24576000 Hz).

In order to quickly convert between timer ticks and nanoseconds,
sched_clock() approximates the conversion using a multiplication and a bit
shift. This results in a small loss of precision and creates an illusion of
nanosecond-level precision that doesn't exist because the timer doesn't
tick fast enough for that.

Since Tensor AIO accumulates windows of counter samples, the sched_clock()
precision loss can also accumulate as a result, leading to inaccurate
results when processing the accumulated statistics (e.g., when calculating
the CPU's real frequency).

Read the system timer directly and compute time in terms of ticks for the
best possible timekeeping precision. Ticks are converted to nanoseconds
losslessly as well, rather than approximating nanoseconds with a bit shift.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>